build: addresses comments
