+++ packages/shared/wallet/src/network-selector/network-selector.tsx
@@ -5,15 +5,7 @@ import { NotionalTheme } from '@notional-finance/styles';
 import { ArrowIcon } from '@notional-finance/icons';
 import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import { Caption, H4, H5, LabelValue, Paragraph } from '@notional-finance/mui';
-import {
-  useTheme,
-  Box,
-  Button,
-  styled,
-  Popover,
-  SxProps,
-  Drawer,
-} from '@mui/material';
+import { useTheme, Box, Button, styled, Popover, SxProps } from '@mui/material';
 import {
   PRODUCTS,
   SupportedNetworks,
@@ -158,169 +150,6 @@ export function PortfolioNetworkSelector({
       onNetworkChange={(network) => portfolioStore.setNetwork(network)}
       hideOnMobile
     />
-  );
-}
-
-export function MobileNetworkSelector() {
-  const selectedNetwork = useSelectedNetwork();
-  const walletBalances = useAccountNetWorth();
-  const portfolioStore = usePortfolioStore();
-
-  return (
-    <NetworkSelectorDrawer
-      availableNetworks={SupportedNetworks}
-      selectedNetwork={selectedNetwork}
-      walletBalances={walletBalances}
-      isPortfolio
-      onNetworkChange={(network) => portfolioStore.setNetwork(network)}
-    />
-  );
-}
-
-function NetworkSelectorDrawer({
-  selectedNetwork,
-  availableNetworks,
-  walletBalances,
-  hideNetWorth,
-  sx,
-  onNetworkChange,
-}: {
-  onNetworkChange?: (network: Network) => void;
-  selectedNetwork?: Network;
-  availableNetworks: Network[];
-  walletBalances: Record<Network, TokenBalance>;
-  hideNetWorth?: boolean;
-  isPortfolio?: boolean;
-  sx?: SxProps;
-}) {
-  const theme = useTheme();
-  const navigate = useNavigate();
-  const { pathname } = useLocation();
-  const userSettings = getFromLocalStorage('userSettings');
-  const [isOpen, setIsOpen] = useState(false);
-  const canSelect = availableNetworks.length > 1;
-
-  const handleClose = (network?: Network) => {
-    setIsOpen(false);
-    if (network && !pathname.includes(network) && selectedNetwork) {
-      if (onNetworkChange && network) onNetworkChange(network);
-      setInLocalStorage('userSettings', { ...userSettings, network: network });
-      navigate(pathname.replace(selectedNetwork, network));
-    }
-  };
-
-  return (
-    <NetworkSelectorWrapper sx={{ ...sx }}>
-      <DropdownButton
-        id="basic-button"
-        aria-controls={isOpen ? 'basic-menu' : undefined}
-        variant="outlined"
-        aria-expanded={isOpen ? 'true' : undefined}
-        disabled={!canSelect}
-        onClick={() => setIsOpen(!isOpen)}
-        startIcon={
-          <TokenIcon
-            symbol={getNetworkSymbol(selectedNetwork)}
-            size={'medium'}
-          />
-        }
-        endIcon={
-          canSelect ? (
-            <Box
-              sx={{
-                borderRadius: '50%',
-                background: theme.palette.info.light,
-                height: theme.spacing(2),
-                width: theme.spacing(2),
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-              }}
-            >
-              <ArrowIcon
-                sx={{
-                  transform: isOpen ? 'rotate(0deg)' : 'rotate(-180deg)',
-                  transition: '.5s ease',
-                  width: theme.spacing(1.5),
-                  color: theme.palette.secondary.light,
-                }}
-              />
-            </Box>
-          ) : undefined
-        }
-        sx={{
-          boxShadow: 'none',
-          padding: '6px 12px',
-          borderRadius: '50px',
-          border: theme.shape.borderStandard,
-          color: theme.palette.typography.main,
-          '&:hover': {
-            boxShadow: 'none',
-          },
-          '.MuiButton-endIcon': {
-            marginLeft: '0px',
-          },
-          '.MuiButton-startIcon': {
-            marginStart: '0px',
-          },
-        }}
-      />
-      <Drawer
-        anchor="top"
-        open={isOpen}
-        onClose={() => handleClose()}
-        sx={{
-          '& .MuiDrawer-paper': {
-            height: 'fit-content',
-            borderBottomLeftRadius: theme.shape.borderRadius(),
-            borderBottomRightRadius: theme.shape.borderRadius(),
-            boxShadow: theme.shape.shadowLarge(),
-            border: theme.shape.borderStandard,
-            background: theme.palette.background.default,
-          },
-        }}
-      >
-        <NetworkInnerWrapper
-          hideNetWorth={hideNetWorth}
-          theme={theme}
-          sx={{ marginTop: '72px', padding: '16px !important' }}
-        >
-          <Box sx={{ padding: theme.spacing(3), paddingLeft: '0px' }}>
-            <H4 sx={{ fontWeight: 700, color: theme.palette.typography.light }}>
-              <FormattedMessage defaultMessage={'NETWORK'} />
-            </H4>
-          </Box>
-          <Box
-            sx={{
-              margin: 'auto',
-              maxHeight: 'calc(50vh - 80px)', // Account for header
-              overflowY: 'auto',
-            }}
-          >
-            {availableNetworks.map((n, i) => (
-              <NetworkSelectorButton
-                key={n}
-                isLast={i === availableNetworks.length - 1}
-                isSelected={n === selectedNetwork}
-                network={n}
-                handleClick={() => handleClose(n)}
-                balance={walletBalances[n]}
-                hideNetWorth={true}
-                showCheck
-                sx={{
-                  borderRadius: theme.shape.borderRadius(),
-                  marginBottom: theme.spacing(2),
-                  background:
-                    n === selectedNetwork
-                      ? theme.palette.info.light
-                      : theme.palette.background.paper,
-                }}
-              />
-            ))}
-          </Box>
-        </NetworkInnerWrapper>
-      </Drawer>
-    </NetworkSelectorWrapper>
   );
 }
